refactor: Health 컴포넌트 이벤트 시스템 및 데이터 타입 변경
- Health 컴포넌트의 OnDamaged 이벤트를 ScriptableObject(GameEventSO)에서 C# Action/delegate 방식으로 리팩토링했습니다.
  - 기존 방식은 다수의 Health 인스턴스(예: 여러 적)가 있을 때, 하나의 이벤트에 모든 인스턴스가 불필요하게 반응하는 문제를 야기했습니다.
  - 이제 각 인스턴스가 개별적으로 이벤트를 처리하도록 개선했습니다.

- 체력(HP) 관련 데이터 타입을 float에서 int로 변경했습니다.
  - UI에 체력을 표시할 때 float으로 인해 발생하던 근사값 출력 오류를 해결하고, 데이터 무결성을 확보하기 위함입니다.
  - Health, IHealthInfo, HUDController 등 체력을 다루는 모든 코드에 일관되게 적용됩니다.

+++ Assets/_Game/Scripts/UI/HUDController.cs
@@ -77,14 +77,14 @@ public class HUDController : MonoBehaviour
     {
         if (_health == null || _hpFill == null) return;
 
-        float max = Mathf.Max(1, _health.MaxHP);
-        float cur = Mathf.Clamp(_health.CurrentHP, 0, max);
-        _hpFill.fillAmount = cur / max;
+        int max = Mathf.Max(1, _health.MaxHP);
+        int cur = Mathf.Clamp(_health.CurrentHP, 0, max);
+        _hpFill.fillAmount = (float)cur / (float)max;
 
         if (_hpText != null)
         {
             _sb.Clear();
-            _sb.Append("HP ").Append((int)cur).Append(" / ").Append((int)max);
+            _sb.Append("HP ").Append(cur).Append(" / ").Append(max);
             _hpText.text = _sb.ToString();
         }
     }
